feat: filter in scan result
This PR separates filter from detection information in the ScanResult
structure. As a consequence of this new design we now have information
about the matching filters in the Filter structure. It is useful to
debug filter or simply to display which filter rule matched. Filter
structure have its own tags member allowing to include tag information
in filtering rules to bring context to the filter or group filters by
family.