perf: increase scan throughput
This PR get rid of string field cloning to gain around 10% scan speed. As a consequence
some lifetime parameters needed to be introduced in Event and FieldGetter traits
but this should be transparent to end users using the derive macros.
For manual trait implementation this will need to be fixed.