fix: dependency rule selectivity issue
A bug occurred when a dependency rule was more selective than the rule it was
used in. The dependency rule incorrectly attempted to match fields even if the
rule couldn't apply to the event (due to the `match-on` directive). This caused
the embedding rule to fail if the dependency rule encountered an error.